Formatted output changed to reflect power usage.

+++ robot/src/main/java/com/team4/robot/PowerManager.java
@@ -2,6 +2,6 @@ package com.team4.robot;
 
 interface PowerManager {
     void recharge();
-    void updateBatteryLevel(int units);
-    int getBatteryLevel();
+    void updateBatteryLevel(double units);
+    double getBatteryLevel();
 }
